backend/requirements/base.txt

# Django
django==2.0.3
djangorestframework==3.8.2
django_filter==1.1.0
dj_database_url==0.4.2

psycopg2-binary==2.7.4

# Raven
raven==5.22.0

# Misc
python-decouple==3.1
requests==2.18.4
